+++ python/pytests/conftest.py
@@ -30,7 +30,13 @@ class GoldenFixture(object):
         filename = self._filename("jsonl")
 
         if self._save:
-            df.to_json(filename, orient="records", lines=True, date_format="iso", date_unit="ns")
+            df.to_json(
+                filename,
+                orient="records",
+                lines=True,
+                date_format="iso",
+                date_unit="ns",
+            )
 
         golden = pd.read_json(
             filename,
@@ -110,19 +116,3 @@ def golden(
         ), f"golden directory {dirname} does not exist. run with `--save-golden` to create it."
 
     return GoldenFixture(dirname, test_name, save)
-
-
-@pytest.fixture
-def source() -> kd.sources.CsvString:
-    content = "\n".join(
-        [
-            "time,key,m,n",
-            "1996-12-19T16:39:57,A,5,10",
-            "1996-12-19T16:39:58,B,24,3",
-            "1996-12-19T16:39:59,A,17,6",
-            "1996-12-19T16:40:00,A,,9",
-            "1996-12-19T16:40:01,A,12,",
-            "1996-12-19T16:40:02,A,,",
-        ]
-    )
-    return kd.sources.CsvString(content, time_column_name="time", key_column_name="key")
