+++ python/pytests/golden/mean_test/test_mean_unwindowed.jsonl
@@ -1,6 +1,6 @@
-{"_time":"1996-12-19T16:39:57.000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"mean_m":5.0,"n":10.0,"mean_n":10.0}
-{"_time":"1996-12-19T16:39:58.000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"mean_m":24.0,"n":3.0,"mean_n":3.0}
-{"_time":"1996-12-19T16:39:59.000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"mean_m":11.0,"n":6.0,"mean_n":8.0}
-{"_time":"1996-12-19T16:40:00.000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"mean_m":11.0,"n":9.0,"mean_n":8.3333333333}
-{"_time":"1996-12-19T16:40:01.000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"mean_m":11.3333333333,"n":null,"mean_n":8.3333333333}
-{"_time":"1996-12-19T16:40:02.000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"mean_m":11.3333333333,"n":null,"mean_n":8.3333333333}
+{"_time":"1996-12-19T16:39:57.000000000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"mean_m":5.0,"n":10.0,"mean_n":10.0}
+{"_time":"1996-12-19T16:39:58.000000000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"mean_m":24.0,"n":3.0,"mean_n":3.0}
+{"_time":"1996-12-19T16:39:59.000000000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"mean_m":11.0,"n":6.0,"mean_n":8.0}
+{"_time":"1996-12-19T16:40:00.000000000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"mean_m":11.0,"n":9.0,"mean_n":8.3333333333}
+{"_time":"1996-12-19T16:40:01.000000000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"mean_m":11.3333333333,"n":null,"mean_n":8.3333333333}
+{"_time":"1996-12-19T16:40:02.000000000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"mean_m":11.3333333333,"n":null,"mean_n":8.3333333333}
