+++ python/pytests/golden/null_test/test_is_null.jsonl
@@ -1,6 +1,6 @@
-{"_time":"1996-12-19T16:39:57.000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"is_null_m":false,"is_not_null_m":true,"n":10.0,"is_null_n":false,"is_not_null_n":true}
-{"_time":"1996-12-19T16:39:58.000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"is_null_m":false,"is_not_null_m":true,"n":3.0,"is_null_n":false,"is_not_null_n":true}
-{"_time":"1996-12-19T16:39:59.000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"is_null_m":false,"is_not_null_m":true,"n":6.0,"is_null_n":false,"is_not_null_n":true}
-{"_time":"1996-12-19T16:40:00.000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"is_null_m":true,"is_not_null_m":false,"n":9.0,"is_null_n":false,"is_not_null_n":true}
-{"_time":"1996-12-19T16:40:01.000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"is_null_m":false,"is_not_null_m":true,"n":null,"is_null_n":true,"is_not_null_n":false}
-{"_time":"1996-12-19T16:40:02.000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"is_null_m":true,"is_not_null_m":false,"n":null,"is_null_n":true,"is_not_null_n":false}
+{"_time":"1996-12-19T16:39:57.000000000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"is_null_m":false,"is_not_null_m":true,"n":10.0,"is_null_n":false,"is_not_null_n":true}
+{"_time":"1996-12-19T16:39:58.000000000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"is_null_m":false,"is_not_null_m":true,"n":3.0,"is_null_n":false,"is_not_null_n":true}
+{"_time":"1996-12-19T16:39:59.000000000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"is_null_m":false,"is_not_null_m":true,"n":6.0,"is_null_n":false,"is_not_null_n":true}
+{"_time":"1996-12-19T16:40:00.000000000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"is_null_m":true,"is_not_null_m":false,"n":9.0,"is_null_n":false,"is_not_null_n":true}
+{"_time":"1996-12-19T16:40:01.000000000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"is_null_m":false,"is_not_null_m":true,"n":null,"is_null_n":true,"is_not_null_n":false}
+{"_time":"1996-12-19T16:40:02.000000000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"is_null_m":true,"is_not_null_m":false,"n":null,"is_null_n":true,"is_not_null_n":false}
