+++ python/pytests/golden/stddev_test/test_stddev_unwindowed.jsonl
@@ -1,6 +1,6 @@
-{"_time":"1996-12-19T16:39:57.000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"stddev_m":null,"n":10.0,"stddev_n":null}
-{"_time":"1996-12-19T16:39:58.000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"stddev_m":null,"n":3.0,"stddev_n":null}
-{"_time":"1996-12-19T16:39:59.000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"stddev_m":6.0,"n":6.0,"stddev_n":2.0}
-{"_time":"1996-12-19T16:40:00.000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"stddev_m":6.0,"n":9.0,"stddev_n":1.6996731712}
-{"_time":"1996-12-19T16:40:01.000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"stddev_m":4.9216076867,"n":null,"stddev_n":1.6996731712}
-{"_time":"1996-12-19T16:40:02.000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"stddev_m":4.9216076867,"n":null,"stddev_n":1.6996731712}
+{"_time":"1996-12-19T16:39:57.000000000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"stddev_m":null,"n":10.0,"stddev_n":null}
+{"_time":"1996-12-19T16:39:58.000000000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"stddev_m":null,"n":3.0,"stddev_n":null}
+{"_time":"1996-12-19T16:39:59.000000000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"stddev_m":6.0,"n":6.0,"stddev_n":2.0}
+{"_time":"1996-12-19T16:40:00.000000000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"stddev_m":6.0,"n":9.0,"stddev_n":1.6996731712}
+{"_time":"1996-12-19T16:40:01.000000000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"stddev_m":4.9216076867,"n":null,"stddev_n":1.6996731712}
+{"_time":"1996-12-19T16:40:02.000000000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"stddev_m":4.9216076867,"n":null,"stddev_n":1.6996731712}
