+++ python/pytests/golden/variance_test/test_variance_unwindowed.jsonl
@@ -1,6 +1,6 @@
-{"_time":"1996-12-19T16:39:57.000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"variance_m":null,"n":10.0,"variance_n":null}
-{"_time":"1996-12-19T16:39:58.000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"variance_m":null,"n":3.0,"variance_n":null}
-{"_time":"1996-12-19T16:39:59.000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"variance_m":36.0,"n":6.0,"variance_n":4.0}
-{"_time":"1996-12-19T16:40:00.000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"variance_m":36.0,"n":9.0,"variance_n":2.8888888889}
-{"_time":"1996-12-19T16:40:01.000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"variance_m":24.2222222222,"n":null,"variance_n":2.8888888889}
-{"_time":"1996-12-19T16:40:02.000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"variance_m":24.2222222222,"n":null,"variance_n":2.8888888889}
+{"_time":"1996-12-19T16:39:57.000000000","_subsort":0,"_key_hash":12960666915911099378,"_key":"A","m":5.0,"variance_m":null,"n":10.0,"variance_n":null}
+{"_time":"1996-12-19T16:39:58.000000000","_subsort":1,"_key_hash":2867199309159137213,"_key":"B","m":24.0,"variance_m":null,"n":3.0,"variance_n":null}
+{"_time":"1996-12-19T16:39:59.000000000","_subsort":2,"_key_hash":12960666915911099378,"_key":"A","m":17.0,"variance_m":36.0,"n":6.0,"variance_n":4.0}
+{"_time":"1996-12-19T16:40:00.000000000","_subsort":3,"_key_hash":12960666915911099378,"_key":"A","m":null,"variance_m":36.0,"n":9.0,"variance_n":2.8888888889}
+{"_time":"1996-12-19T16:40:01.000000000","_subsort":4,"_key_hash":12960666915911099378,"_key":"A","m":12.0,"variance_m":24.2222222222,"n":null,"variance_n":2.8888888889}
+{"_time":"1996-12-19T16:40:02.000000000","_subsort":5,"_key_hash":12960666915911099378,"_key":"A","m":null,"variance_m":24.2222222222,"n":null,"variance_n":2.8888888889}
